Allow to use different prime field element implementations
Before this change, even though the `F: PrimeField` was generic in
`Poseidon` struct and traits it implements, Poseidon parameters were
still defined as `Fq`.

To allow usage of different prime field element types (i.e. `Fr`), the
parameters have to be generic as well.

However, since the `new()` method is not a part of `PrimeField` trait,
functions returining parameters can't be generic, so they are handled
with a macro instead.

Signed-off-by: Michal Rostecki <vadorovsky@gmail.com>