Change params to be compatible with circomlib (#5)
Our previous parameters were compatible with arkworks-gadgets, but not with circomlib. Since we are using circom and circomlib, the compatibility with them is more important.

This PR also adds an xtask for converting parameters generated by hadeshash to Rust source code, so we will be able to add more different parameters if needed.