Define and enforce number of inputs separately from width
Previously when the number of inputs was lower than width - 1, we were
filling up the state of hasher with zeros up to the width, which is
one of the recommended methods of padding described in section 4.2 of
Poseidon hash paper[0].

However, doing so inside the library without user's interference is not
the best thing to do and might lead to misusing the API. Users might not
be aware of the fact they provided less inputs and then might take the
unexpected result for granted.

Also, there are different padding strategies.

The solution to make it more safe here is:

* Defining *width* and *number of inputs* separately. That makes
  potential padding explicit.

To make the use of Rust type system, we are switching from storing
parameters like width, alpha and round numbers in the
`PoseidonParameters` struct to providing them as const generics. That
way we can enforce the particular input size on compile time, without
having to handle or return custom errors on runtime.

[0] https://eprint.iacr.org/2019/458.pdf