fix(core): Enforce strict length check for hashed bytes
Ensure that each slice of bytes, provided as an input in
`PoseidonBytesHasher`, has a size which matches the byte length of
the modulus of the prime field, calculated as:

```
(MODULUS_BIT_SIZE + 7) / 8
```

Where modulus bit size is different for each prime field.

The following methods from `ark_ff`, used usually for converting
byte sequences to prime field elements, perform this check, but
instead of returning an error, they simply take a subslice of the
input, potentially leading to collisions:

* `ark_ff::PrimeField::from_be_bytes_mod_order`
* `ark_ff::PrimeField::from_le_bytes_mod_order`
* `ark_ff::Field::from_random_bytes`

`ark_ff` doesn't provide any method which would perform such check
with returning a recoverable error.