test: Check hashes from byte slice inputs with lower length (#38)
We allow to pass byte slices with lower length than the one indicated
by the modulus of prime field. To ensure that it works, provide a
test which compares that hashes produced from smaller inputs are
equivalent to hashes produces from bigger inputs, filled up with
zeros.

Such inputs with different sizes are still representing the same
prime field elements.